fix: correct GroupSizeEncoding numInGroup type fallback from uint to ushort
The GetNumInGroupType() method in MessagesCodeGenerator returned 'uint' (uint32)
as fallback instead of 'ushort' (uint16), which is the SBE spec default for
numInGroup. This caused CS0266 compilation errors in generated code for any
schema defining numInGroup as uint16.

Changes:
- MessagesCodeGenerator.GetNumInGroupType(): fallback from 'uint' to 'ushort'
- GroupDefinition.NumInGroupType: default from 'uint' to 'ushort'
- Re-enabled 5 benchmark files (removed .disabled suffix)
- Added unit tests for both uint16 and uint32 numInGroup schemas
- Updated KNOWN_ISSUES.md to reflect the fix

+++ src/SbeCodeGenerator/Generators/MessagesCodeGenerator.cs
@@ -302,14 +302,13 @@ namespace SbeSourceGenerator.Generators
         private static string GetNumInGroupType(string dimensionType, SchemaContext context)
         {
             // Look up the type of the numInGroup field in the dimension composite type
-            // Default to "uint" if not found for backward compatibility
             var key = $"{dimensionType}.numInGroup";
             if (context.CompositeFieldTypes.TryGetValue(key, out string? numInGroupType))
             {
                 return numInGroupType;
             }
-            // Fallback to "uint" for backward compatibility with existing schemas
-            return "uint";
+            // Fallback to "ushort" — the SBE spec default for numInGroup is uint16
+            return "ushort";
         }
 
         private static string RegisterGeneratedTypeName(SchemaContext context, string originalName)
